+++ effectful-core/CHANGELOG.md
@@ -1,3 +1,7 @@
+# effectful-core-2.5.0.0 (????-??-??)
+* Add `plusEff` (specialized version of `<|>`) to `Effectful.NonDet` and make
+  `emptyEff` and `sumEff` generate better call stacks.
+
 # effectful-core-2.4.0.0 (2024-10-08)
 * Add utility functions for handling effects that take the effect handler as the
   last parameter to `Effectful.Dispatch.Dynamic`.